+++ scripts/package-scripts/create-new-model-folder.ts
@@ -129,9 +129,9 @@ const createPackageJSON = (name: string, description: string, keywords = DEFAULT
     }
   },
   "scripts": {
-    "scaffold:dependencies": `ts-node ../../scripts/package-scripts/scaffold-dependencies.ts --src models/${name} --config models/scaffolder.ts`,
+    "scaffold:dependencies": `ts-node ../../scripts/package-scripts/scaffold-dependencies.cts --src models/${name} --config models/scaffolder.cts`,
     "lint:fix": "pnpm lint --fix",
-    "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.js src --ext .ts",
+    "lint": "pnpm scaffold:dependencies && eslint -c ../.eslintrc.cjs src --ext .ts",
     "prepublishOnly": "pnpm lint && pnpm build && pnpm validate:build",
     "validate:build": `ts-node ../../scripts/package-scripts/validate-build.ts models/${name}`,
     "build": `ts-node ../../scripts/package-scripts/build-model.ts ${name} -o cjs -o esm -o umd`
