+++ scripts/package-scripts/scaffold-dependencies.cts
